refactor(rust): incorporate review comments
