+++ implementations/rust/ockam/ockam_api/src/config/mod.rs
@@ -1,10 +1,6 @@
-mod boolean_var;
-mod duration_var;
 mod level_var;
 pub mod lookup;
 mod url_var;
 
-pub use boolean_var::*;
-pub use duration_var::*;
 pub use level_var::*;
 pub use url_var::*;
