+++ implementations/rust/ockam/ockam_api/src/logs/span_exporters.rs
@@ -11,7 +11,6 @@ struct DecoratedSpanExporter<S: SpanExporter> {
 #[async_trait]
 impl<S: SpanExporter> SpanExporter for DecoratedSpanExporter<S> {
     fn export(&mut self, batch: Vec<SpanData>) -> BoxFuture<'static, ExportResult> {
-        // debug!("exporting {} spans", batch.len());
         self.exporter.export(batch)
     }
