+++ implementations/rust/ockam/ockam_command/src/lib.rs
@@ -19,6 +19,7 @@
 mod admin;
 mod arguments;
 mod authority;
+mod command;
 mod command_events;
 mod command_global_opts;
 mod completion;
@@ -26,6 +27,7 @@ mod configuration;
 mod credential;
 mod docs;
 pub mod enroll;
+pub mod entry_point;
 mod environment;
 pub mod error;
 mod flow_control;
@@ -37,9 +39,6 @@ mod manpages;
 mod markdown;
 mod message;
 pub mod node;
-mod ockam_command;
-pub mod ockam_run;
-mod ockam_subcommand;
 mod operation;
 mod output;
 pub mod pager;
@@ -56,6 +55,7 @@ pub mod shutdown;
 mod sidecar;
 mod space;
 mod status;
+mod subcommand;
 mod subscription;
 pub mod tcp;
 mod terminal;
@@ -66,11 +66,11 @@ mod version;
 mod worker;
 
 pub use arguments::*;
+pub use command::*;
 pub use command_events::*;
 pub use command_global_opts::*;
 pub use error::*;
 pub use global_args::*;
-pub use ockam_command::*;
-pub use ockam_subcommand::*;
 pub use pager::*;
+pub use subcommand::*;
 pub use terminal::*;
